+++ ft_strlcat.c
@@ -6,17 +6,17 @@
 /*   By: guiferre <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/29 19:41:07 by guiferre          #+#    #+#             */
-/*   Updated: 2024/11/08 03:21:24 by guiferre         ###   ########.fr       */
+/*   Updated: 2024/11/14 16:47:14 by guiferre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-	unsigned int	i;
-	unsigned int	sum;
-	unsigned int	dsize;
-	unsigned int	ssize;
+	size_t	i;
+	size_t	sum;
+	size_t	dsize;
+	size_t	ssize;
 
 	dsize = ft_strlen(dest);
 	ssize = ft_strlen(src);
